tls: handle network exceptions

+++ include/fluent-bit/flb_tls.h
@@ -27,8 +27,8 @@
 
 int io_tls_write(struct flb_thread *th, struct flb_output_plugin *out,
                  void *data, size_t len, size_t *out_len);
-ssize_t io_tls_read(struct flb_thread *th, struct flb_output_plugin *out,
-                    void *buf, size_t len);
+int io_tls_read(struct flb_thread *th, struct flb_output_plugin *out,
+                void *buf, size_t len);
 
 int flb_io_tls_connect(struct flb_output_plugin *out,
                        struct flb_thread *th,
